refactor(CheckoutPage): CheckoutPage 컴포넌트 책임 분리
